Added MqLocationStatic
