+++ packages/mq_prayer_time/lib/src/location/location.dart
@@ -1,5 +1,7 @@
+export 'client/mq_location_client.dart';
 export 'service/mq_location_service.dart';
 export 'service/mq_location_service_impl.dart';
+export 'static/mq_location_static.dart';
 export 'storage/mq_location_storage.dart';
 export 'storage/mq_location_storage_impl.dart';
 export 'storage/mq_location_storage_mock.dart';
